chore: Separate exports to avoid bundling analytics metadata utilities
