+++ package.json
@@ -14,6 +14,8 @@
     ".": "./index.js",
     "./dom": "./dom/index.js",
     "./internal": "./internal/index.js",
+    "./internal/analytics-metadata": "./internal/analytics-metadata/index.js",
+    "./internal/analytics-metadata/utils": "./internal/analytics-metadata/utils.js",
     "./package.json": "./package.json"
   },
   "repository": {
