+++ src/internal/analytics-metadata/__tests__/utils.test.tsx
@@ -3,7 +3,7 @@
 
 import React from 'react';
 import { render } from '@testing-library/react';
-import { getGeneratedAnalyticsMetadata } from '../';
+import { getGeneratedAnalyticsMetadata } from '../utils';
 import { METADATA_ATTRIBUTE } from '../attributes';
 import { ComponentOne, ComponentThree } from './components';
